requirements.txt
requests~=2.32.3
langchain-community~=0.3.15
langchain~=0.3.15
langsmith~=0.2.11
chromadb~=0.5.23
ollama~=0.4.5
numpy~=1.26.4
scikit-learn~=1.6.1
pymongo~=4.13.0
langchain-openai~=0.3.18
python-dotenv~=1.0.1
pymupdf4llm~=0.0.17
nltk~=3.9.1
langchain-text-splitters~=0.3.5
kaggle~=1.6.17
streamlit~=1.41.1
PyYAML~=6.0.2
tqdm~=4.67.1